requirements.txt
torch>=2.0.0
transformers>=4.30.0
datasets>=2.14.0
Pillow>=9.0.0
pandas>=1.5.0
matplotlib>=3.5.0
numpy>=1.23.0
nltk>=3.8.0
rouge-score>=0.1.2
scikit-learn>=1.2.0
tqdm>=4.65.0
requests>=2.28.0
accelerate>=0.20.0
sentencepiece>=0.1.99
accelerate>=0.20.0
